Add files via upload
Incorporated a mobile navigation and also used JQuery to add the same sticky class to it when scrolling past the landing page. Used Jquery to slideToggle when clicking on the mobile-nav-icon and to remove its icon class and add in a cross icon class, else go back to the hamburger icon.